+++ reikna/cluda/ocl.py
@@ -32,13 +32,6 @@ class Thread(api_base.Thread):
         else:
             return ValueError("The value provided is not Device, Context or CommandQueue")
 
-    def supports_dtype(self, dtype):
-        if dtypes.is_double(dtype):
-            extensions = self._context.devices[0].extensions
-            return "cl_khr_fp64" in extensions or "cl_amd_fp64" in extensions
-        else:
-            return True
-
     def allocate(self, size):
         return cl.Buffer(self._context, cl.mem_flags.READ_WRITE, size=size)
 
@@ -61,14 +54,16 @@ class Thread(api_base.Thread):
     def synchronize(self):
         self._queue.finish()
 
-    def _compile(self, src):
-        options = "-cl-mad-enable -cl-fast-relaxed-math" if self._fast_math else ""
+    def _compile(self, src, fast_math=False):
+        options = "-cl-mad-enable -cl-fast-relaxed-math" if fast_math else ""
         return cl.Program(self._context, src).build(options=options)
 
 
 class DeviceParameters:
 
     def __init__(self, device):
+
+        self._device = device
 
         if device.platform.name == 'Apple' and device.type == cl.device_type.CPU:
         # Apple is being funny again.
@@ -108,6 +103,13 @@ class DeviceParameters:
 
         self.min_mem_coalesce_width = {4: 16, 8: 16, 16: 8}
         self.local_mem_size = device.local_mem_size
+
+    def supports_dtype(self, dtype):
+        if dtypes.is_double(dtype):
+            extensions = self._device.extensions
+            return "cl_khr_fp64" in extensions or "cl_amd_fp64" in extensions
+        else:
+            return True
 
 
 class Kernel(api_base.Kernel):