+++ reikna/core/computation.py
@@ -180,20 +180,22 @@ class Computation:
     def _translate_tree(self, translator):
         return self._tr_tree.translate(translator)
 
-    def _get_plan(self, tr_tree, translator, thread):
-        plan_factory = lambda: ComputationPlan(tr_tree, translator, thread)
+    def _get_plan(self, tr_tree, translator, thread, fast_math):
+        plan_factory = lambda: ComputationPlan(tr_tree, translator, thread, fast_math)
         args = [
             KernelArgument(param.name, param.annotation.type)
             for param in tr_tree.get_root_parameters()]
         return self._build_plan(plan_factory, thread.device_params, *args)
 
-    def compile(self, thread):
+    def compile(self, thread, fast_math=False):
         """
         Compiles the computation with the given :py:class:`~reikna.cluda.api.Thread` object
         and returns a :py:class:`~reikna.core.computation.ComputationCallable` object.
+        If ``fast_math`` is enabled, the compilation of all kernels is performed using
+        the compiler options for fast and imprecise mathematical functions.
         """
         translator = Translator.identity()
-        return self._get_plan(self._tr_tree, translator, thread).finalize()
+        return self._get_plan(self._tr_tree, translator, thread, fast_math).finalize()
 
     def _build_plan(self, plan_factory, device_params, *args):
         """
@@ -247,12 +249,13 @@ class ComputationPlan:
     Computation plan recorder.
     """
 
-    def __init__(self, tr_tree, translator, thread):
+    def __init__(self, tr_tree, translator, thread, fast_math):
         """__init__()""" # hide the signature from Sphinx
 
         self._thread = thread
         self._tr_tree = tr_tree
         self._translator = translator
+        self._fast_math = fast_math
 
         self._nested_comp_idgen = IdGen('_nested')
         self._persistent_value_idgen = IdGen('_value')
@@ -408,7 +411,8 @@ class ComputationPlan:
         kernel = self._thread.compile_static(
             template_def, kernel_name, global_size, local_size=local_size,
             render_args=[kernel_declaration] + kernel_argobjects,
-            render_kwds=render_kwds)
+            render_kwds=render_kwds,
+            fast_math=self._fast_math)
 
         self._kernels.append(PlannedKernelCall(kernel, kernel_leaf_names, adhoc_values))
 
@@ -431,7 +435,8 @@ class ComputationPlan:
         new_tree = computation._translate_tree(translator)
         new_tree.reconnect(self._tr_tree)
 
-        self._append_plan(computation._get_plan(new_tree, translator, self._thread))
+        self._append_plan(computation._get_plan(
+            new_tree, translator, self._thread, self._fast_math))
 
     def _append_plan(self, plan):
         self._kernels += plan._kernels