+++ reikna/fft.py
@@ -215,7 +215,7 @@ def get_common_kwds(dtype, device_params):
         min_blocks=helpers.min_blocks,
         mul=functions.mul(dtype, dtype),
         polar=functions.polar(dtypes.real_for(dtype)),
-        cdivs=functions.div(dtype, dtypes.real_for(dtype)))
+        cdivs=functions.div(dtype, numpy.uint32, out_dtype=dtype))
 
 
 class LocalFFTKernel: